fix header and footer tags
